refactor(knowledge): relocate relationRankingLogic into relations (#209)
