+++ src/architecture/knowledge/relations/relationRankingLogic.ts
